feat: 优先读取 materialConfig & 合并 type/info.js
