+++ tools/ice-devtools/lib/generate.js
@@ -9,16 +9,19 @@ module.exports = function generate(cwd) {
   const pkgJson = pkg.getPkgJSON(cwd);
   const { materialConfig } = pkgJson;
 
+  // 目前 materialConfig 不需要实际字段
   if (!materialConfig) {
     logger.fatal(message.invalid);
   }
 
   const { type } = materialConfig;
+  // 生成的物料名称，type 属于兼容字段
+  const filename = type ? `${type}-materials` : 'materials';
 
   generateMaterialsDatabases(
-    type ? `${type}-materials` : 'materials',
+    pkgJson.name,
     cwd,
-    materialConfig
+    filename,
   ).then(() => {
     console.log(chalk.cyan('Success! materials db generated'));
     console.log();
