+++ tools/ice-devtools/utils/type.js
@@ -7,6 +7,13 @@ const pkgJSON = require('./pkg-json');
  */
 function getType(workdir) {
   const json = pkgJSON.getPkgJSON(workdir);
+
+  // 优先读取 materialConfig
+  if (json.materialConfig && json.materialConfig.type) {
+    return json.materialConfig.type;
+  }
+
+  // 兼容逻辑
   if ('blockConfig' in json) {
     return 'block';
   }
